// frontend/src/shared/api/generated/employers/models/vacancyStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Benefit Employers
 * OpenAPI spec version: 0.1.0
 */

export type VacancyStatus = typeof VacancyStatus[keyof typeof VacancyStatus];


export const VacancyStatus = {
  draft: 'draft',
  published: 'published',
  closed: 'closed',
} as const;